LayoutTests/http/tests/loading/resourceLoadStatistics/request-storage-access-same-origin-sandboxed-iframe-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
frame "theIframe" - didStartProvisionalLoadForFrame
main frame - didFinishDocumentLoadForFrame
frame "theIframe" - didCommitLoadForFrame
frame "theIframe" - didFinishDocumentLoadForFrame
CONSOLE MESSAGE: line 66: PASS document.hasStorageAccess was granted.
frame "theIframe" - didHandleOnloadEventsForFrame
main frame - didHandleOnloadEventsForFrame
frame "theIframe" - didFinishLoadForFrame
main frame - didFinishLoadForFrame
Tests that same-origin iframe storage access is granted if the iframe is sandboxed and has the allow token.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
